Change Workflow API. Introduced new parameter desiredCommandFrequency.
* (Incompatible change) Introduced new property named
desiredCommandFrequency that stands for setting for commands frequency
(between 2 and 15)
* Bump version to 0.3.0-SNAPSHOT

+++ src/main/java/org/obd/metrics/api/GenericWorkflow.java
@@ -17,7 +17,7 @@ import lombok.extern.slf4j.Slf4j;
 final class GenericWorkflow extends AbstractWorkflow {
 
 	GenericWorkflow(PidSpec pidSpec, String equationEngine, ReplyObserver observer,
-	        Lifecycle lifecycle, Long commandFrequency) throws IOException {
+	        Lifecycle lifecycle, Integer commandFrequency) throws IOException {
 		super(pidSpec, equationEngine, observer, lifecycle, commandFrequency);
 	}
